+++ services/locations/src/components/App/Dashboard/Location/GenerateQRCodes/QrPrint/QrPrint.styled.js
@@ -1,0 +1,40 @@
+import styled from 'styled-components';
+
+export const QrPrintWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 24px 0;
+  width: 100%;
+  border-top: 1px solid rgb(151, 151, 151);
+  background-color: rgba(248, 249, 250);
+  margin-bottom: -32px;
+`;
+
+export const QrPrintLogo = styled.img`
+  height: 16px;
+  width: 110px;
+  margin-left: 32px;
+`;
+
+export const LinkWrapper = styled.a`
+  display: flex;
+  justify-content: flex-end;
+  margin-right: 32px;
+  width: auto;
+  cursor: pointer;
+  text-decoration: none;
+`;
+
+export const QrPrintText = styled.div`
+  margin: 16px 0 24px 32px;
+  font-family: Montserrat-Medium, sans-serif;
+  font-size: 14px;
+  font-weight: 500;
+`;
+
+export const PrintLink = styled.div`
+  font-family: Montserrat-Bold, sans-serif;
+  font-size: 14px;
+  font-weight: bold;
+  text-transform: uppercase;
+`;