+++ services/locations/src/constants/links.js
@@ -1,4 +1,6 @@
 export const TERMS_CONDITIONS_LINK =
   'https://luca-app.de/app-terms-and-conditions/';
 
+export const QR_PRINT_LINK = 'https://www.qr-print.de/luca';
+
 export const WEB_APP_BASE_PATH = `https://${window.location.host}/webapp/`;