+++ services/scanner/src/components/QrScanner/ScanForm/ScanForm.react.js
@@ -94,6 +94,8 @@ export const ScanForm = ({ scanner }) => {
       setIsSuccess,
       checkForAdditionalData,
     });
+
+    setInputValue('');
   };
 
   const handleChange = event => {